refactor: dto 패키지 위치 변경 및 Article update 영속성 사용으로 변경
