+++ src/main/java/techcourse/myblog/controller/AccountController.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.controller.dto.UserDto;
+import techcourse.myblog.dto.UserDto;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.service.AccountService;
