+++ src/main/java/techcourse/myblog/domain/Article.java
@@ -1,6 +1,7 @@
 package techcourse.myblog.domain;
 
 import lombok.*;
+import techcourse.myblog.dto.ArticleDto;
 
 import javax.persistence.*;
 
@@ -8,6 +9,7 @@ import javax.persistence.*;
 @Getter
 @ToString
 @EqualsAndHashCode(of = {"id"})
+@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class Article {
     @Id
@@ -26,15 +28,13 @@ public class Article {
         this.author = author;
     }
 
-    public Article(Long id, String title, String contents, String coverUrl, User author) {
-        this.id = id;
-        this.title = title;
-        this.contents = contents;
-        this.coverUrl = coverUrl;
-        this.author = author;
-    }
-
     public boolean isNotMatchAuthor(User target) {
         return author.getId() != target.getId();
+    }
+
+    public void update(ArticleDto articleDto) {
+        this.title = articleDto.getTitle();
+        this.contents = articleDto.getContents();
+        this.coverUrl = articleDto.getCoverUrl();
     }
 }
