+++ src/main/java/techcourse/myblog/dto/CommentDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.controller.dto;
+package techcourse.myblog.dto;
 
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
